Module halo2_ecc::fields::fp

Structs

FpChip

Reduced

Wrapper around FieldPoint to guarantee this is a “reduced” representation of an Fp field element. A reduced representation guarantees that there is a unique representation of each field element. Typically this means Uints that are less than the modulus.

Type Definitions

BaseFieldChip

FpConfig